+++ routes/delete/deleteRoom.js
@@ -1,14 +1,15 @@
 const express = require('express');
-const path = require('path');
-const readRoom = require('../../db/read/readRoom');
-const deleteRoom = require('../../db/delete/deleteRoom');
 const router = express.Router();
+const path = require('path');
+const deleteRoom = require('../../db/delete/deleteRoom');
+const readUser = require('../../db/read/readUser');
 const validSession = require('../functions/userSession');
 const dbFile = path.join(__dirname, '../../db/database.db');
 
 router.post('/', (req, res) => {
     const user = validSession(req.session);
-    if (req.session.validSession) {
+    const userPrivilege = readUser(dbFile, user).userPrivilege;
+    if (req.session.validSession && userPrivilege == 'administrator') {
         deleteRoom(dbFile, req.body.id);
         res.redirect('/editRooms');
     } else {
